refactor(codegen): split data-enum DTO coercion into its own module
enums.rs reached 856 lines once the coercion logic landed, over the 800-line
file-modularization threshold. Move the coercion concern — PYO3_DTO_COERCE_HELPER,
CoercibleShape / coercible_payload, coercible_field_init,
data_enum_needs_dto_coercion, pyo3_wire_schema_const_name — into a cohesive
`dto_coercion` module; enums.rs keeps variant-constructor emission. Public
re-export paths are unchanged and generated output is byte-identical. enums.rs
856 -> 656; dto_coercion.rs 222.
