+++ src/codegen/generators/mod.rs
@@ -1,6 +1,7 @@
 use ahash::AHashMap;
 
 pub mod binding_helpers;
+pub mod dto_coercion;
 pub mod enums;
 pub mod functions;
 pub mod methods;
@@ -145,12 +146,15 @@ pub use binding_helpers::{
     gen_named_let_bindings_pub, gen_named_let_bindings_with_augmented, gen_serde_let_bindings, gen_unimplemented_body,
     has_named_params, is_simple_non_opaque_param, wrap_return, wrap_return_with_mutex, wrap_return_with_mutex_mapped,
 };
-pub use enums::{
-    CoercibleShape, PYO3_DTO_COERCE_HELPER, coercible_payload, data_enum_needs_dto_coercion, enum_has_data_variants,
-    gen_enum, gen_pyo3_data_enum, gen_pyo3_data_enum_with_coercion, gen_pyo3_data_enum_with_mapper,
+pub use dto_coercion::{
+    CoercibleShape, PYO3_DTO_COERCE_HELPER, coercible_payload, data_enum_needs_dto_coercion,
     pyo3_wire_schema_const_name,
 };
 pub(crate) use enums::{collect_variant_constructors, variant_field_init};
+pub use enums::{
+    enum_has_data_variants, gen_enum, gen_pyo3_data_enum, gen_pyo3_data_enum_with_coercion,
+    gen_pyo3_data_enum_with_mapper,
+};
 pub use functions::{
     collect_explicit_core_imports, collect_trait_imports, gen_function, gen_function_with_mutex,
     has_unresolved_trait_methods,
